// src/dtos/edition-dto.ts
export interface EditionDTO {
  title?: string;
  publish_date?: string;
  publishers?: string[];
  description?: string;
  number_of_pages?: number;
  avg_number_of_pages?: number;
  cover_edition_key?: string;
  cover_i?: number;
  isbn_10?: string[];
  isbn_13?: string[];
  subjects?: string[];
}